backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
asyncio==4.0.0
attrs==25.3.0
bitarray==3.7.1
cached-property==2.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.1
coverage==7.10.6
crypto_cpp_py==1.4.5
cytoolz==1.0.1
Django==5.2.6
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
ecdsa==0.18.0
eth-account==0.13.7
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-tester==0.13.0b1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lark==1.2.2
lru-dict==1.3.0
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow_dataclass==8.6.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parsimonious==0.10.0
pillow==11.3.0
poseidon_py==0.1.5
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
py-ecc==8.0.0
py-evm==0.12.1b1
py-geth==6.2.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rlp==4.1.0
rpds-py==0.27.1
safe-pysha3==1.0.5
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
starknet-py==0.19.0
swagger-spec-validator==3.0.4
sympy==1.12.1
toolz==1.0.0
trie==3.1.0
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
web3==7.13.0
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
dj-database-url==3.0.1
